+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 5.1.0-beta.1
 
 - Allow functions in `GROUP BY` clause of SOQL queries
+- Support Apex bind variables (`:expr`) on the left-hand side of SOQL `INCLUDES`/`EXCLUDES` conditions
+- Support Apex bind variables (`:expr`) in SOSL `WITH DIVISION`/`NETWORK`/`PRICEBOOKID`/`METADATA` clauses
 
 ## 5.0.0 - 2026-04-21